+++ console/src/components/Services/RemoteSchema/Relationships/RemoteSchemaRelationRenderer.tsx
@@ -1,6 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useGetAllRemoteSchemaRelationships } from '@/features/MetadataAPI';
 import { RemoteSchemaRelationshipTable } from '@/features/RelationshipsTable';
+import { Button } from '@/new-components/Button';
+import { RiAddCircleFill } from 'react-icons/ri';
+// eslint-disable-next-line no-restricted-imports
+import { RemoteSchemaToDbForm } from '@/features/RemoteRelationships/RemoteSchemaRelationships/components/RemoteSchemaToDB';
 
 type RemoteSchemaRelationRendererProp = {
   remoteSchemaName: string;
@@ -14,20 +18,39 @@ export const RemoteSchemaRelationRenderer = ({
     isLoading,
     isError,
   } = useGetAllRemoteSchemaRelationships();
-  if (isError) {
-    return <div>Error in fetching remote schema relationships.</div>;
-  }
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-  if (!remoteSchemaRels || !remoteSchemaRels.length) {
-    return <div>No remote schema relationships found!</div>;
-  }
+  const [isFormOpen, setIsFormOpen] = useState(false);
+
   return (
-    <RemoteSchemaRelationshipTable
-      remoteSchemaRels={remoteSchemaRels}
-      showActionCell={false}
-      remoteSchema={remoteSchemaName}
-    />
+    <>
+      {isError ? (
+        <div>Error in fetching remote schema relationships.</div>
+      ) : isLoading ? (
+        <div>Fetching remote schema relationships...</div>
+      ) : (
+        <RemoteSchemaRelationshipTable
+          remoteSchemaRels={remoteSchemaRels ?? []}
+          showActionCell={false}
+          remoteSchema={remoteSchemaName}
+        />
+      )}
+
+      {isFormOpen ? (
+        <RemoteSchemaToDbForm
+          sourceRemoteSchema={remoteSchemaName}
+          closeHandler={() => setIsFormOpen(!isFormOpen)}
+          onSuccess={() => setIsFormOpen(false)}
+        />
+      ) : (
+        <Button
+          icon={<RiAddCircleFill />}
+          onClick={() => {
+            setIsFormOpen(!isFormOpen);
+          }}
+          data-test="add-a-new-rs-relationship"
+        >
+          Add a new relationship
+        </Button>
+      )}
+    </>
   );
 };