+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - server: refactor GQL query static analysis and improve OpenAPI warning messages
 - console: add support for setting aggregation query permissions for ms sql server
+- console: add RS-to-DB (only postgres & citus) relationships feature to remote schemas tab
 - cli: fix remote schema metadata formatting issues (#7608)
 - cli: fix query collections metadata formatting issues (#7616)